+++ test/test-data/one-test.ts
@@ -1,0 +1,12 @@
+import { EntityCache, StaleEntities } from "gql-cache";
+import { CachePatch } from "../../src";
+
+export interface OneTest {
+  readonly name: string;
+  readonly only?: true;
+  readonly patches: ReadonlyArray<CachePatch>;
+  readonly cacheBefore: EntityCache;
+  readonly cacheAfter: EntityCache;
+  readonly staleBefore?: StaleEntities;
+  readonly staleAfter?: StaleEntities;
+}
